Added new capabilities -----------------------
Added two new function "AllocPool" and "FreePool". These can be used to allocate a memory pool which could be used to allocate
memory for a fixed size resource.

-----Function syntax-----
SYS_ERROR AllocPool(size_t resource_size, void **buffer, uint64_t *pool_id);
SYS_ERROR FreePool(void *buffer, uint64_t *pool_id);

The pool_id is a unique value which is used to add and free resources in a particular pool.
To allocate a resource to a particular pool, it's unique pool_id must be provided.
FreePool frees the resource within the memory pool. If the memory pool doesn't have any more
resources then it automatically deletes the entire pool and changes the pool id to 0.
To specify that you want to create a new pool just pass a pool id of 0 to AllocPool function.

Fixed makefile bugs
----------------------

The makefiles had a bug which allowed a directory ending with 's' or 'c' to be represented as normal .c or .s file and
passed to the compiler. This obviously resulted in compilation and linker error. This bug has now been cleared.

Added resource directory
-----------------------

The resource directory contains a bunch of .psf files which are the only font file format RTOS presently understands.
The default font used is "zap-light18.psf", but you can change that within efi_main function in boot.c file in boot directory.
Just change the name to one of the other font names in the directory to let RTOS use that. Make sure to prepend "\\" before the name.