Changes
Changed the way AllocMem works. Now it will always give a PAGESIZE aligned pointer. This is because of the removal of the malloc
like interface.

New

Added functions such as add_free_mem_entry and add_allocated_mem_entry to surpass memory protocols and add a memory
free list or remove a memory free list from memory. But the caller should use it carefully, as if the added free member
or allocated member is already in list, the results would be unpreditable. These functions are used to add more regions
of memory into the allocator's list which was not mapped in at initialisation of memory manager.